Tests: more explicit file and function names in consistence tests
